Add workflow for requesting job scrapes.
This gives the frontend scrape request functionality, so clicking on
a job will automatically queue a job and request a scrape from the
API server, automatically refreshing the status until it is finished.
Once support exists in StackViz for loading from a UUID URL, this
can be updated to redirect the user automatically.