Refactor the request_scrape task
This refactors the request_scrape task. First, the task now uses
easily-pluggable scanner functions to allow new functionality to be
added easily in the future. Additionally, errors are handled more
consistently, and a new 'message' field is included (and is visible to
clients) so that errors both occur more often when they should
(instead of 'successfully' returning junk data), and explain the issue
when they do occur.

Note that this adds modifications to the database structure, so old
tables should be dropped before updating.